Found and fixed bug for large dimension Rules-based co-kriging. Finding coefficients of linear model with interactions used more evaluations than specified Aux Solver, leading to undefined behaviour. Now if this is the case, aux solver simply does two passes of AMOEBA solver (and nothing else).

+++ cpp_code/surrogate_models.hpp
@@ -474,7 +474,7 @@ class AdaptiveCoKriging: public CoKriging{
 
 		vector<VectorXd> points_;
 		vector<double> values_;
-		double modelFunctionDimension_;
+		int modelFunctionDimension_;
 	};
